+++ src/layer/domain/router/model/eav/value/fetch.ts
@@ -4,13 +4,13 @@ import { standardizeUrl } from '../../../../../data/model/domain/url';
 export class FetchResponse {
   constructor(
     private readonly xhr: XMLHttpRequest,
-    private readonly redirect: boolean,
+    private readonly redirectable: boolean,
   ) {
     assert(this.xhr instanceof XMLHttpRequest);
     assert(this.document instanceof Document);
     void Object.freeze(this);
   }
-  public readonly url = this.redirect && this.xhr.responseURL
+  public readonly url = this.redirectable && this.xhr.responseURL
     ? standardizeUrl(this.xhr.responseURL)
     : '';
   public readonly header = (name: string): string | null =>